Update TopK type to int64 and upgrade goai to v0.13.0
Refactor the TopK field in configuration structs and logic from int to int64 for consistency and compatibility. Upgrade the goai library to version 0.13.0 and integrate its new configuration options, including TopP, Temperature, and TopK parameters.

+++ internal/config/config.go
@@ -46,7 +46,7 @@ type LLMConfig struct {
 	Streaming   bool    `yaml:"streaming"`
 	TopP        float64 `yaml:"top_p"`
 	Temperature float64 `yaml:"temperature"`
-	TopK        int     `yaml:"top_k"`
+	TopK        int64   `yaml:"top_k"`
 }
 
 // FrontendConfig represents the frontend configuration
